requirements.txt
# Deep learning
torch>=2.0.0
torchvision>=0.15.0

# Medical / CV
opencv-python-headless>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0

# API serving
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
gradio>=4.0.0

# Experiment tracking
mlflow>=2.8.0
wandb>=0.16.0

# Data versioning
dvc>=3.30.0

# Utilities
pandas>=2.0.0
matplotlib>=3.7.0
pydantic>=2.0.0

# Testing & code quality
pytest>=7.4.0
ruff>=0.1.0
black==25.12.0

# Middleware
medical-ai-middleware @ git+https://github.com/moebouassida/medical-ai-middleware.git
requests>=2.31.0
boto3>=1.34.0

# HuggingFace deployment
huggingface_hub>=0.19.0
